requirements.txt
python-decouple==3.8
sqlalchemy==2.0.32
pyodbc==5.2.0